@@ -1,32 +1,62 @@
 package com.fullstack.inventory.dto;
 
-import javax.validation.constraints.Min;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.Min;
+import jakarta.validation.constraints.NotNull;
 
 public class InventoryRequest {
     
-    @NotNull(message = "Quantity is required")
-    @Min(value = 0, message = "Quantity must be positive")
-    private Integer quantity;
+    @NotNull(message = "Product code cannot be null")
+    private String productCode;
     
-    @Min(value = 0, message = "Min stock must be positive")
-    private Integer minStock;
+    @Min(value = 0, message = "Quantity cannot be negative")
+    private int quantity;
     
-    @Min(value = 1, message = "Max stock must be at least 1")
-    private Integer maxStock;
+    @Min(value = 0, message = "Price cannot be negative")
+    private double price;
     
+    @Min(value = 0, message = "Reserved quantity cannot be negative")
+    private int reservedQuantity;
+    
+    // Constructors
     public InventoryRequest() {}
     
-    public InventoryRequest(Integer quantity) {
+    public InventoryRequest(String productCode, int quantity, double price, int reservedQuantity) {
+        this.productCode = productCode;
+        this.quantity = quantity;
+        this.price = price;
+        this.reservedQuantity = reservedQuantity;
+    }
+    
+    // Getters and Setters
+    public String getProductCode() {
+        return productCode;
+    }
+    
+    public void setProductCode(String productCode) {
+        this.productCode = productCode;
+    }
+    
+    public int getQuantity() {
+        return quantity;
+    }
+    
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
     
-    public Integer getQuantity() { return quantity; }
-    public void setQuantity(Integer quantity) { this.quantity = quantity; }
+    public double getPrice() {
+        return price;
+    }
     
-    public Integer getMinStock() { return minStock; }
-    public void setMinStock(Integer minStock) { this.minStock = minStock; }
+    public void setPrice(double price) {
+        this.price = price;
+    }
     
-    public Integer getMaxStock() { return maxStock; }
-    public void setMaxStock(Integer maxStock) { this.maxStock = maxStock; }
+    public int getReservedQuantity() {
+        return reservedQuantity;
+    }
+    
+    public void setReservedQuantity(int reservedQuantity) {
+        this.reservedQuantity = reservedQuantity;
+    }
 }
