@@ -1,95 +1,109 @@
 package com.fullstack.product.controller;
 
 import com.fullstack.product.domain.Product;
-import com.fullstack.product.dto.ApiResponse;
-import com.fullstack.product.dto.ProductRequest;
 import com.fullstack.product.service.ProductService;
-import io.swagger.v3.oas.annotations.Operation;
-import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
+import java.util.List;
+import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/v1/products")
-@Tag(name = "Products", description = "Product Management API")
+@RequestMapping("/api/products")
 public class ProductController {
     
     @Autowired
     private ProductService productService;
     
-    @PostMapping
-    @Operation(summary = "Create a new product")
-    public ResponseEntity<ApiResponse<Product>> createProduct(@Valid @RequestBody ProductRequest request) {
-        try {
-            Product product = productService.createProduct(request);
-            return ResponseEntity.status(HttpStatus.CREATED)
-                .body(ApiResponse.success(product, "Product created successfully"));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ApiResponse.error("Failed to create product: " + e.getMessage()));
-        }
+    @GetMapping
+    public List<Product> getAllProducts() {
+        return productService.getAllProducts();
     }
     
     @GetMapping("/{id}")
-    @Operation(summary = "Get product by ID")
-    public ResponseEntity<ApiResponse<Product>> getProduct(@PathVariable String id) {
-        return productService.getProductById(id)
-            .map(product -> ResponseEntity.ok(ApiResponse.success(product)))
-            .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND)
-                .body(ApiResponse.error("Product not found with id: " + id)));
+    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
+        Optional<Product> product = productService.getProductById(id);
+        return product.map(ResponseEntity::ok)
+                     .orElse(ResponseEntity.notFound().build());
+    }
+    
+    @PostMapping
+    public ResponseEntity<?> createProduct(@Valid @RequestBody Product product) {
+        try {
+            Product savedProduct = productService.createProduct(product);
+            return ResponseEntity.status(HttpStatus.CREATED).body(savedProduct);
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.status(HttpStatus.CONFLICT).body(e.getMessage());
+        } catch (Exception e) {
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Error creating product");
+        }
     }
     
     @PutMapping("/{id}")
-    @Operation(summary = "Update product by ID")
-    public ResponseEntity<ApiResponse<Product>> updateProduct(
-            @PathVariable String id, 
-            @Valid @RequestBody ProductRequest request) {
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @Valid @RequestBody Product productDetails) {
         try {
-            return productService.updateProduct(id, request)
-                .map(product -> ResponseEntity.ok(ApiResponse.success(product, "Product updated successfully")))
-                .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(ApiResponse.error("Product not found with id: " + id)));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ApiResponse.error("Failed to update product: " + e.getMessage()));
+            Product updatedProduct = productService.updateProduct(id, productDetails);
+            return ResponseEntity.ok(updatedProduct);
+        } catch (RuntimeException e) {
+            return ResponseEntity.notFound().build();
         }
     }
     
     @DeleteMapping("/{id}")
-    @Operation(summary = "Delete product by ID")
-    public ResponseEntity<ApiResponse<Void>> deleteProduct(@PathVariable String id) {
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
         try {
-            if (productService.deleteProduct(id)) {
-                return ResponseEntity.ok(ApiResponse.success(null, "Product deleted successfully"));
-            } else {
-                return ResponseEntity.status(HttpStatus.NOT_FOUND)
-                    .body(ApiResponse.error("Product not found with id: " + id));
-            }
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ApiResponse.error("Failed to delete product: " + e.getMessage()));
+            productService.deleteProduct(id);
+            return ResponseEntity.ok().build();
+        } catch (RuntimeException e) {
+            return ResponseEntity.notFound().build();
         }
     }
     
-    @GetMapping
-    @Operation(summary = "Get all products with pagination")
-    public ResponseEntity<ApiResponse<Page<Product>>> getAllProducts(
-            @RequestParam(defaultValue = "0") int page,
-            @RequestParam(defaultValue = "10") int size) {
+    @PatchMapping("/{id}/stock")
+    public ResponseEntity<?> updateStock(@PathVariable Long id, @RequestParam Integer quantity) {
         try {
-            Pageable pageable = PageRequest.of(page, size);
-            Page<Product> products = productService.getAllProducts(pageable);
-            return ResponseEntity.ok(ApiResponse.success(products));
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR)
-                .body(ApiResponse.error("Failed to retrieve products: " + e.getMessage()));
+            boolean updated = productService.updateProductStock(id, quantity);
+            if (updated) {
+                return ResponseEntity.ok().build();
+            }
+            return ResponseEntity.notFound().build();
+        } catch (IllegalArgumentException e) {
+            return ResponseEntity.badRequest().body(e.getMessage());
         }
+    }
+    
+    @GetMapping("/sku/{sku}")
+    public ResponseEntity<Product> getProductBySku(@PathVariable String sku) {
+        Optional<Product> product = productService.getProductBySku(sku);
+        return product.map(ResponseEntity::ok)
+                     .orElse(ResponseEntity.notFound().build());
+    }
+    
+    @GetMapping("/search/name")
+    public List<Product> searchProductsByName(@RequestParam String name) {
+        return productService.searchProductsByName(name);
+    }
+    
+    @GetMapping("/search/category")
+    public List<Product> getProductsByCategory(@RequestParam String category) {
+        return productService.getProductsByCategory(category);
+    }
+    
+    @GetMapping("/low-stock")
+    public List<Product> getLowStockProducts(@RequestParam(defaultValue = "5") Integer threshold) {
+        return productService.getLowStockProducts(threshold);
+    }
+    
+    @GetMapping("/out-of-stock")
+    public List<Product> getOutOfStockProducts() {
+        return productService.getOutOfStockProducts();
+    }
+    
+    @GetMapping("/in-stock")
+    public List<Product> getProductsInStock() {
+        return productService.getInStockProducts();
     }
 }
